Refactor MySQL expression generation and adjust random integer bounds
- Updated `Randomly.java` to use `Integer.MAX_VALUE` as the upper bound for the method `getPositiveOrZeroNonCachedInteger()` to better fit within the range of an integer.
- Simplified `MySQLToStringVisitor.java` by removing unnecessary import statements.
- Refactored expression generation in `MySQLToStringVisitor.java` to handle different MySQL data types (BOOLEAN, INT, VARCHAR, FLOAT, DOUBLE, DECIMAL) more efficiently and consistently.
- Improved code readability and maintainability by organizing expression generation logic based on data type and expression depth.

+++ src/sqlancer/Randomly.java
@@ -486,7 +486,7 @@ public final class Randomly {
     }
 
     public static long getPositiveOrZeroNonCachedInteger() {
-        return getNextLong(0, Long.MAX_VALUE);
+        return getNextLong(0, Integer.MAX_VALUE);
     }
 
     public static long getNotCachedInteger(int lower, int upper) {
